requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
Werkzeug
python-dotenv
SQLAlchemy
bleach
PyJWT
pytest
pytest-flask